clean up main package
