+++ cmd/picow-led-server/main.go
@@ -80,52 +80,3 @@ func initLogger(debug bool, host string, port uint) {
 
 	slog.Debug("Flags", "debug", debug, "host", host, "port", port)
 }
-
-type serverHandler struct{}
-
-func (*serverHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	defer func() {
-		if r := recover(); r != nil {
-			slog.Error("Recovered", "r", r)
-		}
-	}()
-
-	crw := &serverResponseWriter{
-		ResponseWriter: w,
-		Hijacker:       w.(http.Hijacker),
-	}
-
-	crw.Header().Set("Access-Control-Allow-Origin", "*")
-	crw.Header().Set("Content-Type", "text/plain")
-	http.DefaultServeMux.ServeHTTP(crw, r)
-
-	log := slog.Warn
-
-	if crw.status >= 200 && crw.status < 300 {
-		if crw.Header().Get("Content-Type") == "" {
-			crw.Header().Set("Content-Type", "application/json")
-		}
-
-		log = slog.Info
-	} else if crw.status >= 500 || crw.status == 0 {
-		log = slog.Error
-	}
-
-	log("Request",
-		"status", crw.status,
-		"addr", r.RemoteAddr,
-		"method", r.Method,
-		"url", r.URL,
-	)
-}
-
-type serverResponseWriter struct {
-	http.ResponseWriter
-	http.Hijacker
-	status int
-}
-
-func (crw *serverResponseWriter) WriteHeader(statusCode int) {
-	crw.status = statusCode
-	crw.ResponseWriter.WriteHeader(statusCode)
-}
